Extracted common AngularJS modules.
